Refactor name filtering code
The code that filters by the first letter of the student names was
refactored to improve readability.

The logic that does the filtering was moved out of the loop that prints
the names. This means that numbering of the students is sequential
whether names have been filtered or not.

The filtering code was refactored into its own method to generally
improve readability.