Load students.csv on startup if no file given
The script now loads students.csv by default if no filename is specified
as an argument. This change has a slight side-effect to the way the
program works.

Previously, if no filename was specified, the program started without
loading any students. This scenario is no longer possible, as the
filename will default to students.csv, and the program will attempt to
load that file.

This created a knockon problem. When the program started for the first
time and there was no students.csv file, it was simply exiting as the
specified (default) couldn't be found. This meant that the program could
never be used for the first time.

To get around this problem, the program now continues, creating the file
if it can't be found, instead of exiting.